+++ database/migrations/2021_09_18_000001_add_collection_disable.php
@@ -1,0 +1,33 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\Schema;
+
+class AddCollectionDisable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('collections', function ($table) {
+            $table->integer('disabled')->default(0)->after('max');
+        });
+
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('collections', function ($table) {
+            $table->dropColumn('disabled');
+        });
+
+    }
+}